Added functionality to GameField, used to be later. Reworked Towers, maybe broke something.
