--- legende-von-morki/src/model/affiliation/Affiliation.java
@@ -1,6 +1,0 @@
-package model.affiliation;
-
-public enum Affiliation {
-    JW,
-    KINGDOM
-}
