--- legende-von-morki/src/model/affiliation/Religion.java
@@ -1,7 +1,0 @@
-package model.affiliation;
-
-public enum Religion {
-    HOLY,
-    UNHOLY,
-    MONK
-}
