--- legende-von-morki/src/model/affiliation/Type.java
@@ -1,9 +1,0 @@
-package model.affiliation;
-
-public enum Type {
-    MELEE,
-    RANGED,
-    MAGIC
-}
-
-//TODO use later
