+++ legende-von-morki/src/model/game/Player.java
@@ -1,4 +1,14 @@
 package model.game;
 
+import model.allies.heroes.AHero;
+
 public class Player {
+
+    private int level;
+    private AHero[] roster;
+    //private int[] numberOfTroops;
+
+    public Player() {
+
+    }
 }
